# internal/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for reading and writing Bear notes via the hub.",
        "title": "Salmon Hub API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/sync/status": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the current sync status including last sync times, queue size, and conflict info.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sync"
                ],
                "summary": "Get sync status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.syncStatusResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "descriptive error message"
                }
            }
        },
...
        "api.syncStatusResponse": {
            "type": "object",
            "properties": {
                "conflict_count": {
                    "type": "integer",
                    "example": 0
                },
                "conflict_note_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "initial_sync_complete": {
                    "type": "string",
                    "example": "true"
                },
                "last_push_at": {
                    "type": "string",
                    "example": "2025-01-15T10:25:00Z"
                },
                "last_sync_at": {
                    "type": "string",
                    "example": "2025-01-15T10:30:00Z"
                },
                "queue_size": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Consumer Bearer token. Format: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}